Enable and fix lints
`-Weverything` enables a lot of very helpful lints.
Add it, disable a couple lints that aren't helpful, and then fix all the remaining lints.

+++ include/time.h
@@ -21,5 +21,5 @@ struct time_components {
 bool time_unix_to_components(time_t time, struct time_components* ret);
 time_t time_unix_from_components(struct time_components const* components);
 
-u32 time_month_to_day_of_year(u8 month, bool is_leap_year);
+u16 time_month_to_day_of_year(u8 month, bool is_leap_year);
 bool time_is_leap_year(i32 year);
